+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   JVM system properties. (Requires `hoist-core >= 10.1` to enable this optional feature.)
 * Provided observable getters `XH.viewportSize`, `XH.isPortrait` and `XH.isLandscape` to allow apps
   to react to changes in viewport size and orientation.
+* Added `GridAutosizeMode.MANAGED` to autosize Grid columns on data or `sizingMode` changes, unless
+  the user has manually modified their column widths. It is now the default `GridAutosizeOptions.mode`.
 
 ### 🐞 Bug Fixes